Implement battle functionality and cleaned warnings

+++ DataAccess/Daos/CardDao.cs
@@ -4,7 +4,7 @@ public class CardDao
 {
     public Guid Id { get; set; }
     public string Name { get; set; } 
-    public float Damage { get; set; }
+    public double Damage { get; set; }
     public string ElementType { get; set; }
     public string CardType { get; set; }
